refactor:improve PSBT signature import Redux patterns
